organize scripts for ether control

+++ Kikusui_PowerSupply/src/common.py
@@ -36,22 +36,39 @@ def writelog(logfile, onoff, voltagelim, currentlim, vol=0., cur=0., timeperiod=
         log = ('{:25s} {:3s} {:3s} {:8.3f} {:8.3f} {:8s} {:8s} {:8s} {:8s}\n'.format(nowStr, onoff, 'NO', voltagelim, currentlim, '--------', '--------', '--------', '--------'))
         pass;
     logfile.write(log)
-    return log
+    return nowStr
 
+def writeitem(logfile, date_str, item, state):
+    log = ('{:25s} {:12s} {:5s}'.format(date_str, item, state))
+    logfile.write(log)
+    pass
 
-def openlog(log_dir) :
+def openlog(log_dir, verbose=0) :
     # Open log file
     now = datetime.datetime.now(timezone('UTC'))
     nowStr  = now.strftime('%Y-%m-%d %H:%M:%S-%Z')
     dateStr = now.strftime('%Y-%m-%d')
-    logfilename = '{}/PMX_{}.dat'.format(log_dir, dateStr)
-    if not os.path.isdir(log_dir) : os.mkdir(log_dir)
-    if os.path.exists(logfilename) :
-      logfile = open(logfilename, 'a+')
-    else :
-      logfile = open(logfilename, 'w' )
-      log = '# Date Time-Timezone ON/OFF CheckValues[Y/N] VoltageLimit[V] CurrentLimit[A] Voltage[V] Current[A] powerOn-timeperiod[sec]\n'
-      logfile.write(log)
+    if verbose == 0:
+      logfilename = '{}/PMX_{}.dat'.format(log_dir, dateStr)
+      if not os.path.isdir(log_dir) : os.mkdir(log_dir)
+      if os.path.exists(logfilename) :
+        logfile = open(logfilename, 'a+')
+      else :
+        logfile = open(logfilename, 'w' )
+        log = '# Date Time-Timezone ON/OFF CheckValues[Y/N] VoltageLimit[V] CurrentLimit[A] Voltage[V] Current[A] powerOn-timeperiod[sec]\n'
+        logfile.write(log)
+        pass
+      pass
+    else:
+      logfilename = '{}/items_{}.dat'.format(log_dir, dateStr)
+      if not os.path.isdir(log_dir) : os.mkdir(log_dir)
+      if os.path.exists(logfilename) :
+        logfile = open(logfilename, 'a+')
+      else :
+        logfile = open(logfilename, 'w' )
+        log = '# Date Time-Timezone Item Start/Stop\n'
+        logfile.write(log)
+        pass
       pass
     return logfile;
